--- src/main/java/com/assignment/St.java
@@ -1,7 +1,0 @@
-package com.assignment;
-
-public class St {
-    public static void stud(){
-     System.out.println("student");
-    }
-}
